Add models
This commit adds model functionality, as well as tests for each of the
features added.  In particular, I choose to use NoSQL elements in order
to speed up execution and make implementation easier.  Since we do not
need JOINS on element or submission information, the disadvantages of
NoSQL are greatly reduced.

In addition, this commit refactors the tests and continuous integration
to use Postgresql and configures flake8 to ignore automatically
generated migrations.

**WARNING: This commit requires the use of Postgresql from henceforth**
**NOTE: For local testing to work, the postgresql user needs the
CREATEDB permission**